Move controls to ui/controls A control is a UI type that doesn't depend on any application types or logic to render. Good examples are buttons, shapes, or text boxes.
+ Make some interfaces export default
+ Sub LearnedCard with CardIdentity since revealed was not used
+ Move SpectatorNote to types
+ No functionality changes
