+++ client/src/game/types/SpectatorNote.ts
