+++ client/src/game/ui/ClueEntry.ts
@@ -2,8 +2,8 @@
 
 // Imports
 import Konva from 'konva';
+import FitText from './controls/FitText';
 import drawLayer from './drawLayer';
-import FitText from './FitText';
 import globals from './globals';
 import HanabiCard from './HanabiCard';
 import * as replay from './replay';
