+++ client/src/game/ui/Elements.ts
@@ -1,27 +1,27 @@
 // Imports
 import Konva from 'konva';
 import Suit from '../types/Suit';
-import Arrow from './Arrow';
-import Button from './Button';
 import ButtonGroup from './ButtonGroup';
 import CardLayout from './CardLayout';
-import ClickArea from './ClickArea';
 import ClueLog from './ClueLog';
 import ColorButton from './ColorButton';
-import CurrentPlayerArea from './CurrentPlayerArea';
+import Arrow from './controls/Arrow';
+import Button from './controls/Button';
+import ClickArea from './controls/ClickArea';
+import CurrentPlayerArea from './controls/CurrentPlayerArea';
+import FitText from './controls/FitText';
+import SharedTurnsButton from './controls/SharedTurnsButton';
+import Shuttle from './controls/Shuttle';
+import StrikeSquare from './controls/StrikeSquare';
+import StrikeX from './controls/StrikeX';
+import ThreeLineButton from './controls/ThreeLineButton';
+import TimerDisplay from './controls/TimerDisplay';
 import Deck from './Deck';
-import FitText from './FitText';
 import FullActionLog from './FullActionLog';
 import MultiFitText from './MultiFitText';
 import NameFrame from './NameFrame';
 import PlayStack from './PlayStack';
 import RankButton from './RankButton';
-import SharedTurnsButton from './SharedTurnsButton';
-import Shuttle from './Shuttle';
-import StrikeSquare from './StrikeSquare';
-import StrikeX from './StrikeX';
-import ThreeLineButton from './ThreeLineButton';
-import TimerDisplay from './TimerDisplay';
 
 export default class Elements {
   // The main screen
