--- client/src/game/ui/ImageWithTooltip.ts
@@ -1,7 +1,0 @@
-// Imports
-import Konva from 'konva';
-
-export default class ImageWithTooltip extends Konva.Image {
-  tooltipName: string = '';
-  tooltipContent: string = '';
-}
