--- client/src/game/ui/LearnedCard.ts
@@ -1,8 +1,0 @@
-// Imports
-import Suit from '../types/Suit';
-
-export default interface LearnedCard {
-  suit: Suit | null;
-  rank: number | null;
-  revealed: boolean;
-}
