--- client/src/game/ui/RectWithTooltip.ts
@@ -1,7 +1,0 @@
-// Imports
-import Konva from 'konva';
-
-export default class RectWithTooltip extends Konva.Rect {
-  tooltipName: string = '';
-  tooltipContent: string = '';
-}
