+++ client/src/game/ui/controls/Button.ts
@@ -1,6 +1,6 @@
 // Imports
 import Konva from 'konva';
-import drawLayer from './drawLayer';
+import drawLayer from '../drawLayer';
 import FitText from './FitText';
 
 export default class Button extends Konva.Group {
