+++ client/src/game/ui/controls/ClickArea.ts
