+++ client/src/game/ui/controls/NodeWithTooltip.ts
@@ -1,0 +1,6 @@
+import Konva from 'konva';
+
+export default interface NodeWithTooltip extends Konva.Node {
+  tooltipName?: string;
+  tooltipContent?: string;
+}
