+++ client/src/game/ui/controls/NoteIndicator.ts
