+++ client/src/game/ui/controls/RectWithTooltip.ts
@@ -1,0 +1,8 @@
+// Imports
+import Konva from 'konva';
+import NodeWithTooltip from './NodeWithTooltip';
+
+export default class RectWithTooltip extends Konva.Rect implements NodeWithTooltip {
+  tooltipName: string = '';
+  tooltipContent: string = '';
+}
