+++ client/src/game/ui/controls/StrikeSquare.ts
