+++ client/src/game/ui/controls/TextWithTooltip.ts
@@ -1,6 +1,6 @@
 // Imports
 import Konva from 'konva';
-import { NodeWithTooltip } from './tooltips';
+import NodeWithTooltip from './NodeWithTooltip';
 
 export default class TextWithTooltip extends Konva.Text implements NodeWithTooltip {
   tooltipName: string = '';
