+++ client/src/game/ui/websocket.ts
@@ -11,6 +11,7 @@ import { GameAction, ActionIncludingHypothetical } from '../types/actions';
 import ClueType from '../types/ClueType';
 import Options from '../types/Options';
 import ReplayArrowOrder from '../types/ReplayArrowOrder';
+import SpectatorNote from '../types/SpectatorNote';
 import action from './action';
 import * as arrows from './arrows';
 import { checkLegal } from './clues';
@@ -20,7 +21,6 @@ import * as notes from './notes';
 import pause from './pause';
 import StateObserver from './reactive/StateObserver';
 import * as replay from './replay';
-import SpectatorNote from './SpectatorNote';
 import strikeRecord from './strikeRecord';
 import * as timer from './timer';
 import * as turn from './turn';
